+++ lib/urls.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'urls/finder'
 require_relative 'urls/validator'
 require_relative 'urls/shortener'